set up utils folder for testing
